\name{clearforkporosity}
\alias{clearforkporosity}
\title{Porosity Data}
\description{
Porosity (fraction of void space) from neutron-density, well log for
5,350--5,400 feet below land surface for Permian Age Clear Fork formation,
Ector County, Texas.
}
\usage{
data(clearforkporosity)
}
\format{
A data frame with
\describe{
  \item{POROSITY}{The pre-sorted porosity data.}
}
}
\details{
Although the porosity data was collected at about 1-foot intervals, these
intervals are not provided in the data frame. Further, the porosity data
has been sorted to disrupt the specific depth to porosity relation to remove
the proprietary nature of the original data.
}
\keyword{datasets}
