\name{IRSrefunds.by.state}
\docType{data}
\alias{IRSrefunds.by.state}
\title{U.S.~Internal Revenue Service Refunds by State for Fiscal Year 2006}
\description{
U.S.~Internal Revenue Service refunds by state for fiscal year 2006.
}
\usage{
data(IRSrefunds.by.state)
}
\format{
A data frame with
\describe{
  \item{STATE}{State name.}
  \item{REFUNDS}{Dollars of refunds.}
}
}
\references{
\url{http://www.irs.gov/taxstats/article/0,,id=168593,00.html} accessed in December 2007.
}
\examples{
data(IRSrefunds.by.state)
summary(IRSrefunds.by.state)
}
\keyword{datasets}
