\name{lmom.references}
\alias{lmom.references}
\title{Important References Related to L-moments}
\description{
A subtantial body of statistical research provides foundation for the theory of
L-moments and demonstration of L-moment theory in practice exists.  Whereas, in many ways,
J.R.M. Hosking should be considered the father of L-moments, there are indeed
many contributors to L-moment literature.  Further, R. Serfling and P. Xiao in 2006 have
taken up the reigns of multivariate L-moments. A substantial sampling is provided in the
documentation of this package for the benefit of users.
}
\references{
Asquith, W.H., 1998, Depth-duration frequency of precipitation for Texas: U.S.
Geological Survey Water-Resources Investigations Report 98-4044, 107 p.

Asquith, W.H., 2002, Effects of regulation on L-moments of annual peak
streamflow in Texas: U.S. Geological Survey Water-Resources Investigations
Report 01-4243, 66 p.

Asquith, W.H., 2003, Modeling of runoff-producing rainfall hyetographs in
Texas using L-moment statistics: Ph.D. dissertation, University of Texas at
Austin, Austin, Texas, 386 p.

Asquith, W.H., and Roussel, 2004, Atlas of depth-duration frequency of
precipitation annual maxima for Texas: U.S. Geological Survey Scientific
Investigations Report 2004-5041, 106 p.

Cunnane, C., 1989, Statistical distributions for flood frequency analysis: World
Meteorological Organization Operational Hydrology Report No. 33, variously
paged.

David, H.A., 1981, Order statistics: John Wiley and Sons Inc., New York, 360 p.

Dingman, S.L., 2002, Physical Hydrology: Prentice Hall, New Jersey,
2nd edition, 646 p.

Evans M., Hastings, N.A.J., and Peacock, J. B., 2000, Statistical distributions,
3rd ed.: John Wiley and Sons Inc., 221 p.

Fill, H.D., and Stedinger, J.R., 1995, L-moment and probability plot correlation
coefficient goodness-of-fit tests for the Gumbel distribution and impact of
autocorrelation: Water Resources Research, vol. 31, no. 1, pp. 225-229.

Gilchrist, W.G., 2000, Statistical modeling with quantile functions: Chapman
and Hall CRC Press, Boca Raton, Fla., 320 p.

Greenwood, J.A., Landwehr, J.M., Matalas, N.C., and Wallis, J.R., 1979,
Probability weighted moments-definition and relation to parameters of several
distributions expressible in inverse form: Water Resources Research, vol. 15,
pp. 1029-1054.

Haktanir, T., and Bozduman, A., 1995, A study on sensitivity of the probability-
weighted moments method on the choice of the plotting position formula: Journal
of Hydrology, vol. 168, pp. 265-281.

Hosking, J.R.M., 1986, The theory of probability weighted moments, Research
Report RC12210, IBM Research Division, T.J. Watson Research Center, Yorktown
Heights, New York.

Hosking, J.R.M., 1989, Some theoretical results concerning L-moments, Research
Report RC14492, IBM Research Division, T.J. Watson Research Center, Yorktown
Heights, New York.

Hosking, J.R.M., 1990, L-moments: Analysis and estimation of distributions
using linear combination of order statistics: Journal Royal Statistical
Society, B, vol. 52, no. 1, pp. 105-124.

Hosking, J.R.M., 1992, Moments or L moments? An example comparing two measures
of distributional shape: The American Statistician., vol 46, no. 3, pp. 186-189.

Hosking, J.R.M., 1994, The four-parameter Kappa distribution: IBM Journal of
Research and Development, vol. 38, no. 3, pp. 251-258.

Hosking, J.R.M., 1995, The use of L-moments in the analysis of censored data,
in Recent Advances in Life-Testing and Reliability, edited by N. Balakrishnan,
chapter 29, CRC Press, Boca Raton, Fla., pp. 546-560.

Hosking, J.R.M., 1996, FORTRAN routines for use with the method of L-moments,
version 3: Research Report RC20525, IBM Research Division, T.J. Watson Research
Center, Yorktown Heights, New York.

Hosking, J.R.M., and Wallis, J.R., 1987, Parameter and quantile estimation for
the generalized Pareto distribution: Technometrics, vol. 29, pp. 339-349.

Hosking, J.R.M., and Wallis, J.R., 1993a, Some statistics useful in regional
frequency analysis: Water Resources Research, vol. 29, no. 2, pp. 271-281.

Hosking, J.R.M., and Wallis, J.R., 1993b, Some statistics useful in regional
frequency analysis, Research Report RC-17096, IBM Research Division, T.J.
Watson Research Center, Yorktown Heights, New York, 23 p.

Hosking, J.R.M., and Wallis, J.R., 1995, A comparison of unbiased and plotting-
position estimators of L moments: Water Resources Research, vol. 31, no. 8,
pp. 2019-2025.

Hosking, J.R.M., and Wallis, J.R., 1997, Regional frequency analysis-An
approach based on L-moments: Cambridge University Press, 224 p.

Hosking, J.R.M., Wallis, J.R., and Wood, E.F., 1985, Estimation of the
generalized extreme-value distribution by the method of probability-weighted
moments: Technometrics, vol. 27, pp. 251-261.

Kaigh, W.D., and Driscoll, M.F., 1987, Numerical and graphical data summary
using O-statistics: The American Statistician, vol. 41, no. 1, pp. 25-32.

Karian, Z.A., and Dudewicz, E.J., 2000, Fitting statistical distributions--
The generalized lambda distribution and generalized bootstrap methods: 
CRC Press, Boca Raton, FL, 438 p.

Kirby, W., 1974, Algebraic boundedness of sample statistics: Water Resources
Research, vol. 10, pp. 220-222.

Landwehr, J.M., Matalas, N.C., and Wallis, J.R., 1979a, Probability weighted
moments compared with some traditional techniques in estimating Gumbel
parameters and quantiles: Water Resources Research, vol. 15, no. 5,
pp. 1055-1064.

Landwehr, J.M., Matalas, N.C., and Wallis, J.R., 1979b, Estimation of parameters
and quantiles of Wakeby distributions: Water Resources Research, vol. 15, no. 5,
pp. 1362-1379.

Schaefer, M.G., 1989, Characteristics of extreme precipitation events in
Washington State: Washington State Department of Ecology Report 89-51,
variously paged.

Schaefer, M.G., 1993, Dam safety guidelines technical note 3-Design storm
construction: Washington State Department of Ecology Report 92-55G, variously
paged.

Serfling, R., and Xiao, P., 2006, Multivariate L-moments: preprint

Stedinger, J.R., Vogel, R.M., and Foufoula-Georgiou, E., 1992, Frequency
analysis of extreme events, in Handbook of Hydrology, chapter 18,
editor-in-chief D. A. Maidment: McGraw-Hill, New York.

Vogel, R.M., and Fennessey, N.M., 1993, L moment diagrams should replace
product moment diagrams: Water Resources Research, vol. 29, no. 6, pp. 1745-1752.

Wallis, J.R., Matalas, N.C., and Slack, J.R., 1974, Just a moment!: Water
Resources Research, vol. 10, pp. 211-219.

Wang, Q.J., 1996a, Using partial probability weighted moments to fit the extreme
value distributions to censored samples: Water Resources Research, vol. 32,
no. 6, pp. 1767-1771.

Wang, Q.J., 1996b, Direct sample estimators of L-moments: Water Resources
Research, vol. 32, no. 12., pp. 3617-3619.

Zafirakou-Koulouris, A., Vogel, R.M., Craig, S.M., and Habermeier, J., 1998,
L-moment diagrams for censored observations: Water Resources Research, vol. 34,
no. 5, pp. 1241-1249.
}
\author{ W.H. Asquith}
\keyword{univar}
\keyword{distribution}
