\name{lmomcoBook}
\alias{lmomcoBook}
\title{Show the Errata for Asquith (2011) Book }
\description{
Show the Errata file for the following book

Asquith, W.H., 2011, Distributional analysis with L-moment statistics
       using the R environment for statistical computing: CreateSpace
       Independent Publishing Platform, 2nd printing, ISBN 978-1463508418.
       \url{http://www.amazon.com/Distributional-Statistics-Environment-
       Statistical-Computing/dp/1463508417}

The Errata file is named \code{ERRATA_FOR_ISBN9781463508418.txt} and is available within the \pkg{lmomco} package in the \code{inst} directory.

}
\usage{
lmomcoBook()
}
%\arguments{
%}
\value{
  None.
}
\keyword{documentation}
