\name{lmomcoBook}
\alias{lmomcoBook}
\title{Show the Errata for Asquith (2011) Book }
\description{
Show the Errata file for the following book

Asquith, W.H., 2011, Distributional analysis with L-moment statistics using the R environment for statistical computing: Createspace Independent Publishing Platform, ISBN 978--146350841--8.

Available at \url{https://www.amazon.com} and search for \code{1463508417} (their id number) or the ISBN \cr \code{9781463508418} (do not use hyphens).

The Errata file is named \code{ERRATA_FOR_ISBN9781463508418.txt} and is available within the \pkg{lmomco} package in the \code{inst/} directory.
}
\usage{
lmomcoBook()
}
%\arguments{
%}
\value{
  None.
}
\keyword{documentation}
