\name{lmomcoBook}
\alias{lmomcoBook}
\title{Show the Errata for Asquith (2011) Book }
\description{
Show the Errata file for the following book

Asquith, W.H., 2011, Distributional analysis with L-moment statistics
       using the R environment for statistical computing: CreateSpace
       Independent Publishing Platform, 2nd printing, ISBN 978-1463508418.

Available at \url{http://www.amazon.com} and search for \code{1463508417} (their id number) or the ISBN \cr \code{9781463508418} (do not use hyphens).

The Errata file is named \code{ERRATA_FOR_ISBN9781463508418.txt} and is available within the \pkg{lmomco} package in the \code{inst} directory.

}
\usage{
lmomcoBook()
}
%\arguments{
%}
\value{
  None.
}
\keyword{documentation}
