\encoding{utf8}
\name{.lmomcohash}
\docType{data}
\alias{.lmomcohash}
\title{Storage of Lookup Tables for the lmomco Package}
\description{
This is a hidden data object contained in the \code{R/sysdata.rda} file of the \pkg{lmomco} package.
}
\format{
An \R \code{environment} with \code{data.frame} entries:
\describe{
  \item{AEPkh2lmrTable}{.}
  \item{EMU_lmompara_byeta}{.}
  \item{KMU_lmompara_bykappa}{.}
  \item{RiceTable}{.}
  \item{RiceTable.maxLCV}{.}
  \item{RiceTable.minLCV}{.}
}
}
\keyword{datasets}
